+++ src/gwt/src/org/rstudio/core/client/theme/res/ThemeStyles.java
@@ -156,4 +156,6 @@ public interface ThemeStyles extends CssResource
    String presentationNavigatorLabel();
    
    String notResizable();
+   
+   String dialogTabPanel();
 }